Fix clustering problem.
There is a problem when we meet two files that already have hardlinks
each. It is possible for hardlink.py to idly move a file from one
hardlink to another without ever ultimately saving space. This can
happen on large disks with backup files, stored and occasionally
hardlinked over years. The solution is to always use the file with the
most hardlinks as the source if (stat_info.st_nlink >
temp_stat_info.st_nlink).