Fix test case for match option prefixes
The match option implies matching against filenames, but the actual
check was being done against full paths, which meant to match a file one
always had to begin the match string with an asterisk.  Ie. "*foo" was
required to match a file called "foo".  This adds a test to check for
that bug.