Add test for --filenames-equal iteration order bug
Adds a test in case dir2 is found before dir1.  With the filenames equal
option enabled, the order of iteration could can a false negative (ie.
an improperly passed test).  By changing the initial link relationship,
we can cause the test to fail if the directory iteration order is
reversed.