Protect against old-style verbose options
There is a rare corner case when transitioning from old verbose option
arguments (-v num) to new (-v -v ...), that the OptionParser cannot
catch, when the given 'num' matches an existent directory name.

Add a safeguard to prevent scanning 'num' directories inadvertently when
old-style verbose arguments are still being used.  This *could* block
intentional cases where someone wants to legitimately scan a directory
named as a number.  This seems a rare case, though, and can be worked
around by not specifying the directory name directly after a '-v'
argument (or not using '-v').

An error message is printed to indicate the ambiguous argument, and the
code can be disabled by changing the default global boolean.
Eventually, this safeguard could be removed, when it's safe to assume
that users have transitioned to the new style verbosity arguments.