Rename filename and stat variables for clarity
In hardlink_identical_files(), the work_file_info, temp_filename, and
temp_stat_info variables were confusingly named.  The important
distinction is whether the information was seen previously (and cached),
or is newly discovered, which wasn't properly emphasized.

By prefixing variable names with "cached_" instead of "temp_", the
intent is clearer, especially with all the intermixing of the variable
names in this section.