Fix reporting of space saved by stats and logging
Additional space savings are only made when the destination file of a
hardlink has an nlink count of one.  The file is first unlinked,
lowering its link count to 0, and saving space, before being hardlinked
to (which uses no additional space).  Both the statistics and logging
code were not reporting this properly.