Track how often destination st_nlink drops to zero
Linking two equal files may not save any space, if the removed filename
entry doesn't actually drop an inode nlink count to zero.  When an inode
nlink does drop to zero, and the filename is linked to another inode, we
call this a "consolidation".  We now keep track and report how often
this occurs.