Change filename to pathname where appropriate
Pathnames indicate a full path, meaning a combined dirname and filename,
ie. 'dirname/filename'.  The filename is what is returned by
os.path.basename().

A different times the code is concerned with actual filenames (such as
when matching, or walking the directory tree), so it's helpful to
consistently distinguish between the two.