Test to show that 'anti-clustering' alg is broken
The change that causes linking to always use the file with the greater
nlink as the source, doesn't work, because it can orphan already seen
files, but since only a single pass is made, the files aren't considered
again for linking to the larger cluster.  The problem can be masked with
two clusters, since the there are enough remaining files to allow the
unlinked but already seen files to be relinked.  These tests demonstrate
the issue by using 3 clusters to test with.