Revert "Fix clustering problem."
Go back to always using cached pathname as link() source.

This reverts commit 73d2764e476f7307ba196d0b67c0de338ef55ee7.
This reverts commit 5e9d961212d9d04679bfd3a94a35cbac0c6b15a7.
This reverts commit 591be1ce61af259b119dea45f836092a5e0ae518.

Since this is a single pass algorithm, the link() source should always
be a file that was previously seen and is in the file_hashes cache.  In
this way, the cached file's inode count can only increase, and the
algorithm can link to all equal files (at the cost of doing more link()
calls.)

By reversing the link order when the newly discovered file had a larger
nlink value, files that had already been processed could become unlinked
from the new cluster, and possibly never re-linked to it.  The test case
added previously demonstrates this effect.