Initial checkin of new algorithm
The new algorithm keeps track of filenames and pathnames associated with
each inode, so allows slight performance improvement when used in
equal-filenames mode.  Mainly it's a base for creating an algorithm
which tracks all the possible links before performing them, then sorting
them so that they occur in an optimal order (ie. the minimal number of
link calls).