Move previous hardlink finding out of cached loop
Since we now track all previous inodes, we can determine whether a new
file is already linked to one we've previously seen, without looping
over the file hashes.  This is more accurate, since previously a new
file could be encountered that matched a file's hash, but not be the
same inode as the one in the file_hashes cache, and this wouldn't be
counted as a previously linked file (even though it was linked to a file
that had been encountered).

So now we should at least have an accurate count of the number of
previously linked files.