Break linkify() into multiple generators
Separate the walking of files, from the looking for equal files.
Generate the allowable (ie. matched/non-excluded) files, and feed them
to the file comparison generator.  Add support functions and methods for
collecting and storing per-device inode data, so that linking can be
postponed until after all files are walked.

Also, the new algorithm orders the linking from the inodes with the
highest nlink count, to those with the least nlink counts.  This helps
reduce the amount of re-linking performed, and also reduces the amount
of relinking that is performed when new files are found on subsequent
linking runs.

Finally, it also does some counting of the amount of bytes saved based
on the number of inodes remaining after linking is performed.  Doing the
calculation after all the files data are collected should allow more
accurate accounting.