Remove unneeded sort in linkable_inode_sets
Originally sorting was used to ensure starting on the highest numbered
inode (as the inode values were always less than the key inode number).
However, it's required that a pair of inodes numbers be added to the
linked_inodes in both orderings, ie. (a,b) and (b,a).  This meant the
algorithm could work without having to start at the highest value.

In any case, a heapq could have been used to find the max value, rather
than a full sort.