Add some helpers for creating/updating file_info
There is a lot of file_info unpacking/repacking in the code, so having
some helper functions that can encapsulate it somewhat can be useful.

If support for Python pre-python 2.6 is dropped, we could probably
convert to using NamedTuples for file_info tuples, and possibly save on
some of the unpacking/repacking.