Allow calling Hardlinkable with no options arg
The options parser can now return a default options object that is used
when Hardlinkable is called without being passed an options object.

"hardlinkable" can be imported as a module, and the Hardlinkable class
instantiated and used as a service.  It can generate pairs of linkable
pathnames, or to perform the actual linking (or just getting the stats
on what would have occurred).