Remove chown() after hardlink.
Chown() is a Unix only function, and it's not clear to me that keeping
the uid/gid of the most recently modified file is all that important,
especially since there are restrictions on when the function can
succeed.

It was claimed at the time it was added that it helped with certain
backup solutions, but I think avoiding chown() altogether is worth
reverting the change for now untill a compelling argument for it is
made.