Rename file_hash to inode_hash
We are really calculating a "hash" of inode data, not file data.  This
is to avoid confusion with people/coders thinking the program is doing
SHA hashing (etc) of file contents.

Files are *always* compared byte-for-byte when determining if they are
equal or not (the inode hash is just a quick way to determine files
which *don't* need to have their contents compared)