Move samename logic for fileinfo lookup into call
Makes code intent easier to understand (imo) moving the samename
checking into the call argument.

In general, the filenames-equal logic tends to disrupt the algorithm
logic, as it adds little variants to the normal code-flow of the
algorithm.  Rather than adding a lot of if/else clauses to the
algorithm, I find it simpler to push the samename checks into the helper
functions, allowing the main code to read more easily in the normal (ie.
non-filenames-equal) case.