Add test classes for randomized testing of files
Create tests with directories and files in a randomized lexical order,
to help determine an ordering based bugs (particularly with the samename
option).  Also randomized pre-existing links and file contents somewhat.

We should still be able to determine the correct result in terms of
bytes saved, and links created, and compare with the data gathered in
the LinkingStats objects to make sure they are equal (or the differences
explainable).