Add nlink and path counts to inode stats
We don't always find all the paths to the inodes we discover, because
they may be in other directory trees, or they could be excluded, etc.
So it makes sense to account for them separately (ie. both the inode
nlinks, and the discovered paths) when determining how much space is
"saved".