Add simple file content "checksum" function.
Based on the zlib checksums, mainly because they should be included with
all (most?) Python implementations, and we don't really need strong
cryptographic hashing for this.

We combine the adler32 and crc32 values for a 64-bit value, mainly
because the adler32 (while fast) can produce a limited range of values
when passed a lot of zeros as input.