Store a content checksum for inodes when comparing
Since the file compare will read the content of a file anyway, we can
calculate the content checksum with no real I/O penalty anyway.  We
calculate the checksum on just the first 8K of a file, since that is
what the filecmp module will read anyway.

The checksum can weed out definitely non-equal files, but we still do a
full file comparison to verify a file match, so this technique mainly
helps eliminating long lists of files with the same inode hash (ie.
"size", typically), but different contents.