Remove adler32 from the checksum calculation
Rather than a 64-bit combined checksum (adler and crc), just use the CRC
to save a bit of time and memory; the adler32 likely didn't add any
value.  The main use of the checksums are quickly eliminating definitely
unequal files from the search list, for which 32-bits is more than
sufficient.