Setup stats container storage when quiet selected
Though the verbosity options control when the stats containers store
hardlinks, we also have explicit options for them, which we can set when
the quiet option is used (which resets the verbosity value).

This is mainly helpful when using the Hardlinkable class directly,
because the command itself exits early when there is no output or
actions requested.