Put mismatch stats after file comparison result
The mismatch stats now are collected after a file comparison has been
made and had a true result.  So the mismatches count when an equal file
has been found, despite mismatched stat attributes (possible when
content-only is selected, or certain attributes are ignored).

This gives more meaningful results, as previously the counts could be
highly over-inflated as a result of repeated pre-comparison checks
during the file-by-file search for equal files.